requirements.txt
together
python-dotenv
pydantic==2.7.1
pydantic_core==2.18.2
fastapi
streamlit_chat

uvicorn
requests
gtts
reportlab
streamlit
